Playlist permission

+++ src/main/java/me/crespel/karaplan/service/impl/PlaylistServiceImpl.java
@@ -1,7 +1,10 @@
 package me.crespel.karaplan.service.impl;
 
+import java.util.LinkedHashSet;
 import java.util.Optional;
 import java.util.Set;
+import java.util.UUID;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
@@ -12,11 +15,17 @@ import com.google.common.collect.Sets;
 
 import me.crespel.karaplan.domain.Playlist;
 import me.crespel.karaplan.domain.Song;
+import me.crespel.karaplan.domain.User;
+import me.crespel.karaplan.model.exception.BusinessException;
 import me.crespel.karaplan.repository.PlaylistRepo;
+import me.crespel.karaplan.repository.UserRepo;
 import me.crespel.karaplan.service.PlaylistService;
 
 @Service
 public class PlaylistServiceImpl implements PlaylistService {
+
+	@Autowired
+	protected UserRepo userRepo;
 
 	@Autowired
 	protected PlaylistRepo playlistRepo;
@@ -32,6 +41,23 @@ public class PlaylistServiceImpl implements PlaylistService {
 	}
 
 	@Override
+	public Set<Playlist> findAll(Pageable pageable, User user) {
+		return playlistRepo.findAll(pageable).stream()
+				.map(playlist -> {
+					if (!isMember(user, playlist)) {
+						playlist.setAccessKey(null);
+					}
+					return playlist;
+				})
+				.collect(Collectors.toCollection(LinkedHashSet::new));
+	}
+
+	@Override
+	public Set<Playlist> findAllAuthorized(Pageable pageable, User user) {
+		return Sets.newLinkedHashSet(playlistRepo.findAllByRestrictedOrMembersId(false, user.getId()));
+	}
+
+	@Override
 	public Optional<Playlist> findById(Long id) {
 		return findById(id, false);
 	}
@@ -39,11 +65,33 @@ public class PlaylistServiceImpl implements PlaylistService {
 	@Override
 	@Transactional(readOnly = true)
 	public Optional<Playlist> findById(Long id, boolean includeSongs) {
+		return findById(id, includeSongs, null);
+	}
+
+	@Override
+	@Transactional(readOnly = true)
+	public Optional<Playlist> findById(Long id, boolean includeSongs, User user) {
 		Optional<Playlist> playlist = playlistRepo.findById(id);
-		if (playlist.isPresent() && includeSongs) {
-			playlist.get().getSongs().size(); // Force eager load
+		if (playlist.isPresent()) {
+			if (includeSongs) {
+				playlist.get().getSongs().size(); // Force eager load
+			}
+			if (!isMember(user, playlist.get())) {
+				playlist.get().setAccessKey(null);
+			}
 		}
 		return playlist;
+	}
+
+	@Override
+	@Transactional
+	public Playlist create(String name, User user, boolean restricted) {
+		Playlist playlist = new Playlist().setName(name).setRestricted(restricted);
+		if (restricted) {
+			playlist.setAccessKey(UUID.randomUUID().toString());
+			playlist.getMembers().add(user);
+		}
+		return playlistRepo.save(playlist);
 	}
 
 	@Override
@@ -55,7 +103,10 @@ public class PlaylistServiceImpl implements PlaylistService {
 
 	@Override
 	@Transactional
-	public Playlist addSong(Playlist playlist, Song song) {
+	public Playlist addSong(Playlist playlist, Song song, User user) {
+		if (!isMember(user, playlist)) {
+			throw new BusinessException("User " + user + " is not a member of playlist " + playlist);
+		}
 		playlist.getSongs().add(song);
 		song.getPlaylists().add(playlist);
 		song.updateStats();
@@ -64,7 +115,10 @@ public class PlaylistServiceImpl implements PlaylistService {
 
 	@Override
 	@Transactional
-	public Playlist removeSong(Playlist playlist, Song song) {
+	public Playlist removeSong(Playlist playlist, Song song, User user) {
+		if (!isMember(user, playlist)) {
+			throw new BusinessException("User " + user + " is not a member of playlist " + playlist);
+		}
 		playlist.getSongs().remove(song);
 		song.getPlaylists().remove(playlist);
 		song.updateStats();
@@ -73,8 +127,40 @@ public class PlaylistServiceImpl implements PlaylistService {
 
 	@Override
 	@Transactional
-	public void delete(Playlist playlist) {
+	public Playlist addUser(Playlist playlist, User user, String accessKey) {
+		if (playlist.getRestricted() && playlist.getAccessKey() != null) {
+			if (playlist.getAccessKey().equals(accessKey)) {
+				if (!playlist.getMembers().contains(user)) {
+					playlist.getMembers().add(user);
+					return playlistRepo.save(playlist);
+				}
+			} else {
+				throw new BusinessException("Invalid playlist access key");
+			}
+		}
+		return playlist;
+	}
+
+	@Override
+	@Transactional
+	public void delete(Playlist playlist, User user) {
+		if (!isMember(user, playlist)) {
+			throw new BusinessException("User " + user + " is not a member of playlist " + playlist);
+		}
 		playlistRepo.delete(playlist);
+	}
+
+	@Override
+	public boolean isMember(User user, Playlist playlist) {
+		if (playlist == null) {
+			return false;
+		} else if (user == null) {
+			return true;
+		} else if (!playlist.getRestricted()) {
+			return true;
+		} else {
+			return playlist.getMembers() != null && playlist.getMembers().contains(user);
+		}
 	}
 
 }
